cli: Make render-ignition fragment generation provider-agnostic
Move the Ignition fragment types and rendering from the Azure provider into render_ignition and source platform data through the MetadataProvider trait (admin_username, admin_password_hash), and switch the feature flags to opt-out (--disable-hostname-fragment / --disable-user-fragment).
