+++ dracut/30afterburn/afterburn-ignition-fragment.service
@@ -21,6 +21,6 @@ OnFailureJobMode=isolate
 
 [Service]
 ExecStartPre=/usr/bin/mkdir -p /etc/ignition/base.platform.d/azure
-ExecStart=/usr/bin/afterburn render-ignition --cmdline --render-ignition-dir /etc/ignition/base.platform.d/azure/ --platform-extensions
+ExecStart=/usr/bin/afterburn render-ignition --cmdline --render-ignition-dir /etc/ignition/base.platform.d/azure/
 Type=oneshot
 RemainAfterExit=yes
